Video Editor: Fix for sending garbage content in Fake Buffers
The Video Editor is sending one frame initially to "test" encoder.
This i/p buffer has no data, it is just allocated from the heap and sent to the Encoder.
So, the data this frame contains is garbage which the encoder sometime fails to encode
within the allocated o/p buffer size.
Changes below does a memset on this buffer before sending to Encoder.

Change-Id: I4c8bfa7730f67e36f1e522e1e979382f1413e19c
Signed-off-by: Spurthi <spurthi@ti.com>
Signed-off-by: Sathya Prakash M R <a0393677@ti.com>